LayoutTests/http/tests/preload/dynamic_removing_preload-expected.txt

PASS internals.isPreloaded('../resources/dummy.js'); is false
PASS successfullyParsed is true

TEST COMPLETE
